+++ R/class_gcp_file.R
@@ -66,10 +66,4 @@ store_read_object.tar_gcp_file <- function(store, file) {
   file_move(from = scratch, to = stage)
   stage
 }
-
-#' @export
-store_unload.tar_gcp_file <- function(store, target) {
-  unlink(as.character(target$value$object))
-  NextMethod()
-}
 # nocov end